pyproject.toml
[tool.poetry]
name = "interviewtodolist"
version = "0.1.0"
description = ""
authors = ["Pavel Merkulov <merkulov0pavel@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.1.1"
djangorestframework = "^3.15.2"
celery = "^5.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
